2 - Aug 4

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: aahsan <aahsan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/01 14:23:52 by aahsan            #+#    #+#             */
-/*   Updated: 2022/08/02 10:49:37 by aahsan           ###   ########.fr       */
+/*   Updated: 2022/08/02 13:43:14 by aahsan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 void	*ft_calloc(size_t nmemb, size_t size)
 {
 	void	*ptr;
-	
-	if (size && nmemb > INT_MAX / size)
+
+	if (size && nmemb > SIZE_MAX / size)
 		return (NULL);
 	ptr = (void *) malloc(nmemb * size);
 	if (ptr == NULL)
